(INIT) Setup blog
Prior to this change, the repository was created from the
platen.io template repository. I decided to keep the blog
simple for now, and Platen has some warts that make it
incompatible with newer versions of Hugo.

Rather than resolve those bugs in Platen, I decided to
switch to a very minimal theme. The point of this site
is to write blog posts so I can  get them out of my
head and published. I can always reorganize and update
the blog later or switch to different tooling.

This change:

- Removes all of the irrelevant files from the template.
- Updates the netlify config to use the newest version
  of Hugo.
- Adds headers to Netlify (`static/_headers`) to improve
  security and privacy for readers.
- Adds a (bad) favicon that should be replaced later.
- Sets the hugo theme to `bearcub`.
- Configures the new hugo theme.
- Defines a new readme for the blog.
- Defines GitHub issues for reporting goofs, requesting
  posts, and suggesting theme enhancements.
- Defines a discussion template for new post announcements.
- Defines minimal VS Code configuration.
- Defines minimal gitignore globs.