(MAINT) Fix issue template and config
Prior to this change, the planned maintenance issue template
didn't show up in the issue picker because the YAML defined
`title` as an empty string, which is invalid for the template
schema. Further, the issue template configuration used a
mix of first and third person when referring to me.

This change:

- Removes `title` from the planned maintenace issue template.
- Standardizes on first person for the configuration.