(MAINT) Fix footnote in _On git practices_
Prior to this change, the Markdown for the footnote in _On git practices_
was malformed, causing the first paragraph to appear in the footnote but
placing the second paragraph outside of the footnotes section as an
erroneous note following the sign off at the end of the article.

This change corrects the footnote syntax to indent the lines sufficiently
and I manually tested the rendering to ensure the footnote appears
correctly on the site.