(DOCS) Add footnote to _On git practices_
Prior to this change, there was no way to get to the discussion for the
post or to file a goof for the post from the blog page. I initially
investigated extending the theme to support this, but decided to just
add a footnote for now, given the theme limitations and how rusty my
hugo development skills currently are.

This change:

- Pulls the admonishment not to yell at me out of the note on terms and
  into a paragraph following that note.
- Directs users who want to have a conversation or report a goof to
  check the footnote at the end of the paragraph.
- Defines a new footnote with links to the discussion and a friendly
  link to report a goof, pre-filling the location for the reader.

I want to address this need more robustly in a future change.